Update output folder treatment/ creation
Previously, I defined the output_path as output_path + version already in the config file. However, this is limiting as it does not allow to export output to other versions if required. Therefore, the version is now part of the file.path() statement and not automatically part of the output_path statement.