Fixed that export re-read is not updated
The reading of the exported data was not update even if the cleaning was update. This is because there was no connection between the cleaning and re-reading part. I fixed this by establishing a dependency which is technically not used (there is no new cleaning) but required for the updating.