Fix wrong behaviour in country code and implausible values
For countrycode, I used a coalesce strategy to combine the information of the older variable country_id and countrycode. However, this strategy only works if the variables contain NAs coded as such. But I already recoded the missings according to helpers_missing_values. So, coalesce assigned the values of country_id (as mentioned first) to all data points. Switched to a case_when() strategy that does this correctly.

Even the strategy for classifying implausible values worked, the algorithm overwrote other missing values (replace missing codes with -5). However, to keep the missing coding meaningful this should not happen. The additional conditions take care of this.