Workaround for missing values in mapping tables
Some of the values for modelid, makeid, and emissionstickerid are not present in the mapping tables likely because they are part of previous years. This is a workaround as I replace these values with missings right now. This is has to be updated once we get a response from AS24.