Removed variables from newer_variables that get dropped
Since I drop irrelevant variables before checking newer variables, it led to conflicts if the previously removed variables show up here as newer variables. Therefore, I removed these variables from the list and wrote a test that makes sure that variables that are deleted cannot be in this list of newer variables.